rename pkg 'doubles' to 'testsupport' (#28)
