+++ test/e2e/kubefed_test.go
@@ -3,7 +3,7 @@ package e2e
 import (
 	"context"
 	"github.com/codeready-toolchain/toolchain-common/pkg/cluster"
-	"github.com/codeready-toolchain/toolchain-e2e/doubles"
+	"github.com/codeready-toolchain/toolchain-e2e/testsupport"
 	"github.com/codeready-toolchain/toolchain-e2e/wait"
 	"github.com/operator-framework/operator-sdk/pkg/test"
 	"github.com/stretchr/testify/require"
@@ -15,7 +15,7 @@ import (
 
 func TestKubeFedE2E(t *testing.T) {
 	fedClusterList := &v1beta1.KubeFedClusterList{}
-	ctx, awaitility := doubles.WaitForDeployments(t, fedClusterList)
+	ctx, awaitility := testsupport.WaitForDeployments(t, fedClusterList)
 	defer ctx.Cleanup()
 
 	verifyKubeFedCluster(ctx, awaitility, cluster.Host, awaitility.Member())
@@ -29,7 +29,7 @@ func verifyKubeFedCluster(ctx *test.TestCtx, awaitility *wait.Awaitility, kubeFe
 	current, ok, err := singleAwait.GetKubeFedCluster(kubeFedClusterType, nil)
 	require.NoError(awaitility.T, err)
 	require.True(awaitility.T, ok, "KubeFedCluster should exist")
-	labels := doubles.KubeFedLabels(kubeFedClusterType, current.Labels["namespace"], current.Labels["ownerClusterName"])
+	labels := testsupport.KubeFedLabels(kubeFedClusterType, current.Labels["namespace"], current.Labels["ownerClusterName"])
 
 	awaitility.T.Run("create new KubeFedCluster with correct data and expect to be ready for cluster type "+string(kubeFedClusterType), func(t *testing.T) {
 		// given
@@ -38,7 +38,7 @@ func verifyKubeFedCluster(ctx *test.TestCtx, awaitility *wait.Awaitility, kubeFe
 			current.Spec.APIEndpoint, current.Spec.SecretRef.Name, labels)
 
 		// when
-		err := awaitility.Client.Create(context.TODO(), newFedCluster, doubles.CleanupOptions(ctx))
+		err := awaitility.Client.Create(context.TODO(), newFedCluster, testsupport.CleanupOptions(ctx))
 
 		// then the KubeFedCluster should be ready
 		require.NoError(t, err)
@@ -56,7 +56,7 @@ func verifyKubeFedCluster(ctx *test.TestCtx, awaitility *wait.Awaitility, kubeFe
 			"https://1.2.3.4:8443", current.Spec.SecretRef.Name, labels)
 
 		// when
-		err := awaitility.Client.Create(context.TODO(), newFedCluster, doubles.CleanupOptions(ctx))
+		err := awaitility.Client.Create(context.TODO(), newFedCluster, testsupport.CleanupOptions(ctx))
 
 		// then the KubeFedCluster should be offline
 		require.NoError(t, err)
